Summary: when using torch.fx to symbolic trace nn module, we create Proxy node instead of real Tensor, and Proxy has implemented '__torch_function__'. However, the C++ custom op is just cpp_function. In current cpp_function definition, there is no support for '__torch_function__' override, so we will see complaints about the C++ function expect Tensor input, but got Proxy type. So when calling the cpp_function of the custom op, we first check whether the input args support '__torch_function__' override, if so, we will try to call the override logic instead of the real C++ custom op.

Some perf benchmark:
- sigmoid (input is torch.ones(2))
without change: Op sigmoid had a minimum time of 2.506256103515625 us and a standard deviation of 0.3828478220384568 us.
with change: Op sigmoid had a minimum time of 2.7022361755371094 us and a standard deviation of 0.3112237900495529 us.
- relu (input is torch.ones(2))
without change: Op relu had a minimum time of 2.2428035736083984 us and a standard deviation of 0.4182197153568268 us.
with change: Op relu had a minimum time of 2.4945735931396484 us and a standard deviation of 0.3914815024472773 us.
- cat (input is torch.ones(2), torch.zeros(2))
without change: Op cat had a minimum time of 3.509521484375 us and a standard deviation of 0.5249808309599757 us.
with change: Op cat had a minimum time of 4.532814025878906 us and a standard deviation of 0.6780577823519707 us.

- cat (input is torch.ones(3, 224, 224), torch.zeros(3, 224, 224))
without change: Op cat had a minimum time of 107.0699691772461 us and a standard deviation of 6.263310555368662 us.
with change: Op cat had a minimum time of 106.03904724121094 us and a standard deviation of 5.618293769657612 us.

Test Plan: buck test //caffe2/test:fx -- test_torch_custom_ops

Differential Revision: D23203204