@ezyang 2020-08-28
> Summary: First override torch.ops, then export the custom op as call_function. After symbolic execution, restore the original _Ops to torch.ops.

This PR description needs updating. The patch is pretty involved now and it would be really good to have an actual explanation of what is going on in English.