@ezyang 2020-08-28
Can you explain to me why we have to reimplement the overload resolution procedure to implement `__torch_function__` here? That's pretty unexpected.