@houseroad 2020-08-29
Hi @ezyang, I have updated the summary. In short, custom op is exposed as cpp_function in python, and we don't have specific logic for __torch_function__ override.