@ezyang 2020-08-31
This is probably about as good as it is going to get without a serious rewrite of how python argument parsing works in general. There are some correctness issues I raised in review, please take a look.